% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{add_observed}
\alias{add_observed}
\title{Add observations to data}
\usage{
add_observed(db, dir_obs, variable)
}
\arguments{
\item{db}{tibble of model results.}

\item{dir_obs}{character string of directory with observations.}

\item{variable}{character string of SWAP variable.}
}
\description{
Add observations to data
}
\keyword{internal}
