% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_depth}
\alias{check_depth}
\title{Detect depth information SWAP variable}
\usage{
check_depth(variable)
}
\arguments{
\item{variable}{character string of SWAP variable.}
}
\description{
Detect depth information SWAP variable
}
\keyword{internal}
