% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{check_value_SWAP}
\alias{check_value_SWAP}
\title{Check value of SWAP-variable}
\usage{
check_value_SWAP(value, variable, ...)
}
\arguments{
\item{value}{value of SWAP-variable.}

\item{variable}{character string, name of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Check value of SWAP-variable
}
\keyword{internal}
