% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{copy_SWAP}
\alias{copy_SWAP}
\title{Copy SWAP-model}
\usage{
copy_SWAP(file_swp, dir_run, quiet = FALSE)
}
\arguments{
\item{file_swp}{character string, name of swp-file.}

\item{dir_run}{character string, run-directory to copy \code{file_swp} to.}

\item{quiet}{logical.}
}
\description{
Copy SWAP-model
}
