% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database_SWAP}
\alias{create_database_SWAP}
\title{Create sql-database from xls-file (SWAP-variables)}
\usage{
create_database_SWAP(file_xls, file_sql, ...)
}
\arguments{
\item{file_xls}{character string, name of xls-file.}

\item{file_sql}{character string, name of sql-file.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Create sql-database from xls-file (SWAP-variables)
}
