% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.R
\name{create_meteo_AgERA5}
\alias{create_meteo_AgERA5}
\title{Create meteo files for SWAP (based on AgERA5 data)}
\usage{
create_meteo_AgERA5(dir_out, latitude, longitude, tstart, tend, ...)
}
\arguments{
\item{dir_out}{character string, name of directory}

\item{latitude}{numeric, the latitude of the site in decimal degrees}

\item{longitude}{numeric, the longitude of the site in decimal degrees}

\item{tstart}{date, start of period}

\item{tend}{date, end of period}

\item{...}{further arguments passed to or from other methods}
}
\description{
Based on observations from \href{https://cds.climate.copernicus.eu}{AgERA5}
}
\details{
Download of meteorological data.
Optionally the data is split in yearly files by specifying \code{split} = TRUE.
Missing data can be replaced by specifying a second station with the option \code{replace}.

It is known that the precipitation predicted by AgERA5 is biased in certain areas, particularly in the tropics.
Therefore this procedure provides the opportunity to replace the precipitation estimates from AgERA5 with precipitation
estimates from CHIRPS by specifying \code{opt_rain}.
}
\examples{
# create SWAP meteo files from AgERA5 (do not run!)
#create_meteo_AgERA5(dir_out = "../temp_swap",
# latitude = 52.1, longitude = 5.18, tstart = "2000-01-01", tend = "2001-12-31")
}
