% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.R
\name{create_meteo_SWAP}
\alias{create_meteo_SWAP}
\title{Create meteo files for SWAP (based on KNMI data)}
\usage{
create_meteo_SWAP(dir_out, station, tstart, tend, ...)
}
\arguments{
\item{dir_out}{character string, name of directory}

\item{station}{character string, name(s) of meteo stations}

\item{tstart}{date, start of period}

\item{tend}{date, end of period}

\item{...}{further arguments passed to or from other methods}
}
\description{
Based on observations from \href{https://www.knmi.nl/home}{KNMI}
}
\details{
Download of meteorological data.
Optionally the data is split in yearly files by specifying \code{split} = TRUE.
Missing data can be replaced by specifying a second station with the option \code{replace}.
}
\examples{
# create SWAP meteo files from KNMI (do not run!)
#create_meteo_SWAP(dir_out = "../temp_swap",
# station = "de Bilt", tstart = "1980-01-01", tend = "1980-12-31")
}
