% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{create_SWAP}
\alias{create_SWAP}
\title{Create SWAP-model using sqlite-database}
\usage{
create_SWAP(file_swp, file_sql, run_id, tmplt_swp, dir_met, ...)
}
\arguments{
\item{file_swp}{character string, name of swp-file (to create).}

\item{file_sql}{character string, name of sql-file.}

\item{run_id}{numeric, id of run (sql-datebase \code{file_sql}).}

\item{tmplt_swp}{character string, name of swp-template.}

\item{dir_met}{character string, directory with meteo-file(s).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Create SWAP-model using sqlite-database
}
