% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{delete_depth}
\alias{delete_depth}
\title{Delete depth information from SWAP variable}
\usage{
delete_depth(variable)
}
\arguments{
\item{variable}{character string of SWAP variable.}
}
\description{
Delete depth information from SWAP variable
}
\keyword{internal}
