% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extract_run_id}
\alias{extract_run_id}
\title{Extract swap-files from zip}
\usage{
extract_run_id(run_id, dir_run, dir_out, ...)
}
\arguments{
\item{run_id}{vector, run id.}

\item{dir_run}{character string, directory to extract file.}

\item{dir_out}{character string, directory to extract file.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Extract swap-files from zip
}
