% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{extract_variable_inlist}
\alias{extract_variable_inlist}
\title{Extract output variables of SWAP-model}
\usage{
extract_variable_inlist(file_swp, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Extract output variables of SWAP-model
}
\keyword{internal}
