% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{filter_run_id_SQL}
\alias{filter_run_id_SQL}
\title{Filter run_id from sql-database}
\usage{
filter_run_id_SQL(file_sql, ...)
}
\arguments{
\item{file_sql}{character string, name of sql-file.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Filter run_id from sql-database
}
