% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvG.R
\name{get_cond_MvG}
\alias{get_cond_MvG}
\title{Get conductivity based on pressure head}
\usage{
get_cond_MvG(H, ALPHA, NPAR, LAMBDA, KSAT)
}
\arguments{
\item{H}{pressure head [cm]}

\item{ALPHA}{curve shape parameter [-]}

\item{NPAR}{curve shape parameter [-]}

\item{LAMBDA}{exponent in hydraulic conductivity function [-]}

\item{KSAT}{hydraulic conductivity of saturated soil [cm d-1]}
}
\value{
return conductivity [cm d-1]
}
\description{
Get conductivity based on pressure head
}
\keyword{internal}
