% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_gt}
\alias{get_gt}
\title{Get GT class}
\usage{
get_gt(ghg, glg, opt = "1966")
}
\arguments{
\item{ghg}{value. Average highest groundwaterlevel [cm+mv].}

\item{glg}{value. Average lowest groundwaterlevel [cm+mv].}

\item{opt}{character string. options: 1966, 1977 or 1995.}
}
\value{
return GT based on \code{GHG} and \code{GLG}
}
\description{
Get GT class
}
