% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_gxg}
\alias{get_gxg}
\title{Determine mean highest and mean lowest groundwater level}
\usage{
get_gxg(db, ...)
}
\arguments{
\item{db}{dataframe with columns 'date' and 'value'.}

\item{...}{further arguments passed to or from other methods}
}
\value{
return mean highest and mean lowest groundwater level [cm+mv]
}
\description{
Determine mean highest and mean lowest groundwater level
}
\details{
To derive the mean highest (GHG) and mean lowest (GLG) groundwater level, all values on the 14th and 28th day of the month are selected leaded to 24 values for each hydrologic year (April - March).
From this the three highest (HG3) and three lowest (LG3) values are selected and averaged over a long term period resulting in the GHG and GLG. Normally the long term period should be 8 years at least.

In case of incomplete timeseries of groundwater level observations on the 14th and 28th day of the month a certain shift in days is allowed. As a default a maximum of 5 days is used, but this can be adjusted by \code{max_shift}.
Although the shift in days, it is still possible the timeseries is missing observations. A certain amount of missing observations is allowed (deault 0.9; this can be adjusted by \code{min_frac}.
Furthermore the length of the long term period can be adjusted by \code{min_year} (default: 8 years).

The period over which the groundwater characteristics should be determined can be adjusted by specifying \code{tstart} and/or \code{tend}.
}
