% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_INLIST_SWAP}
\alias{get_INLIST_SWAP}
\title{Extract output variables of SWAP-model}
\usage{
get_INLIST_SWAP(file_swp = NULL, run_info = NULL)
}
\arguments{
\item{file_swp}{character string, name of swp-file (to create).}

\item{run_info}{list, index of run_id.}
}
\description{
Extract output variables of SWAP-model
}
\keyword{internal}
