% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{get_modelperformance}
\alias{get_modelperformance}
\title{Calculate model performance}
\usage{
get_modelperformance(actual, predicted, performance, ...)
}
\arguments{
\item{actual}{numeric vector, observed values}

\item{predicted}{numeric vector, modelled values}

\item{performance}{character vector, model performance to be calculated}

\item{...}{further arguments passed to or from other methods}
}
\description{
Calculate model performance
}
