% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_plot_settings}
\alias{get_plot_settings}
\title{Get plot settings based on SWAP variable}
\usage{
get_plot_settings(variable, ...)
}
\arguments{
\item{variable}{character string of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get plot settings based on SWAP variable
}
\keyword{internal}
