% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_run_id_SQL}
\alias{get_run_id_SQL}
\title{Get run_id from SQL}
\usage{
get_run_id_SQL(file_sql)
}
\arguments{
\item{file_sql}{character string, name of sql-file.}
}
\description{
Get run_id from SQL
}
