% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_run_info_SQL}
\alias{get_run_info_SQL}
\title{Get run_info from SQL}
\usage{
get_run_info_SQL(file_sql, run_id)
}
\arguments{
\item{file_sql}{character string, name of sql-file.}

\item{run_id}{numeric, id of run.}
}
\description{
Get run_info from SQL
}
