% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{get_table_variable_SWAP}
\alias{get_table_variable_SWAP}
\title{Get variable(s) in table to set in SWAP-template}
\usage{
get_table_variable_SWAP(tmplt, table, ...)
}
\arguments{
\item{tmplt}{character string, text of SWAP-template from which names of variable(s) should be extracted.}

\item{table}{character string, name of table in SWAP-template from which names of variables should be extracted.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get variable(s) in table to set in SWAP-template
}
\keyword{internal}
