% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_value_SQL}
\alias{get_value_SQL}
\title{Get value of variable from SQL}
\usage{
get_value_SQL(conn, run_info, variable, ...)
}
\arguments{
\item{conn}{connection to sql-file.}

\item{run_info}{list, index of run_id.}

\item{variable}{character string, name of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get value of variable from SQL
}
