% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{get_value_SWAP}
\alias{get_value_SWAP}
\title{Get value of SWAP-variable}
\usage{
get_value_SWAP(variable, ...)
}
\arguments{
\item{variable}{character string, name of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get value of SWAP-variable
}
\examples{
# specify swp-file
file_swp <- paste0(tempfile(pattern = "swap"), "/swap.swp")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, files = basename(file_swp), exdir = dirname(file_swp))

# extract name of project
get_value_SWAP(file = file_swp, variable = "PROJECT")

# extract number of layers
max(get_value_SWAP(file = file_swp, variable = "SOILPROFILE::ISOILLAY"))

# extract saturated conductivity
get_value_SWAP(file = file_swp, variable = "SOILHYDRFUNC::OSAT")

# clean example
unlink(x = dirname(file_swp), recursive = TRUE)
}
