% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{get_variable_SWAP}
\alias{get_variable_SWAP}
\title{Get variable(s) to set in SWAP-template}
\usage{
get_variable_SWAP(tmplt, ...)
}
\arguments{
\item{tmplt}{character string. text of SWAP-template from which names of variable(s) should be extracted.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Get variable(s) to set in SWAP-template
}
