% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{get_version_SWAP}
\alias{get_version_SWAP}
\title{Get SWAP version}
\usage{
get_version_SWAP(file_swp)
}
\arguments{
\item{file_swp}{character string, name of swp-file.}
}
\description{
Get SWAP version
}
\details{
Extracts SWAP version number from csv-output
}
