% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvG.R
\name{get_wc_h}
\alias{get_wc_h}
\title{Get water content of soil profile at certain depth based on specific pressure head}
\usage{
get_wc_h(file_swp, depth, ...)
}
\arguments{
\item{file_swp}{character string, name of swp-file.}

\item{depth}{character string, range of depth in [cm].}

\item{...}{further arguments passed to or from other methods}
}
\value{
return water content [cm3 cm-3].
}
\description{
Get water content of soil profile at certain depth based on specific pressure head
}
\details{
Pressure head can be specified by \code{h} or \code{pF}.
}
