% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_contour}
\alias{plot_contour}
\title{Create contour-plot}
\usage{
plot_contour(file_swp, variable, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{variable}{character string of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates contour-plot for a specific SWAP variable.
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) and SWAP variable (\code{variable}) a contour-plot is created.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output_tz.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, exdir = dir_out)

# plot contour of water content
plot_contour(file_swp = file_swp, variable = "WC", tstart = "2016-01-01", ylim = c(-200.0, 0))

# clean example
unlink(x = dir_out, recursive = TRUE)
}
