% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cropdev}
\alias{plot_cropdev}
\title{Create timeseries-plot of potential and actual crop development}
\usage{
plot_cropdev(file_swp, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates timeseries-plot of simulated potential and actual crop development
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) a timeseries-plot is created of the potential and actual crop development.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

In case observations are available it is possible to add the observations as red dots in the plot.
A path to a folder with observations should be specified with \code{dir_obs}.
In this folder a csv-file with the name of the SWAP-variable should be present.
The file itself should have two columns: 'DATETIME' and 'VALUE'. The unit of the values should be equal to SWAP.
Optionally the model performance (\code{performance}) can be added in the plot (see also: \link{get_modelperformance}).

By default the crop development of aboveground biomass is plotted for all crops in the crop-calendar during the simulation period.
Specify a list of crops with \code{cropfil} in case only the development of specific crop(s) is preferred.
In case the development of a certain part of the plant is preferred specify \code{growth} (options: aboveground biomass ('AGB'), storage organs ('SO'), leaves ('LV'), stems ('ST'), roots ('RT'), Leaf Area Index ('LAI') or rootzone ('RD').

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, exdir = dir_out)

# plot timeseries of crop development
plot_cropdev(file_swp = file_swp, growth = "AGB")

# plot timeseries of crop development including observations and model performance
plot_cropdev(file_swp = file_swp, growth = "AGB", dir_obs = paste0(dir_out, "/observed"), performance = "rmse")

# clean example
unlink(x = dir_out, recursive = TRUE)
}
