% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gwlevel}
\alias{plot_gwlevel}
\title{Create timeseries-plot of simulated groundwater level}
\usage{
plot_gwlevel(file_swp, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates timeseries-plot of groundwater level simulated by SWAP.
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) a timeseries-plot is created of the simulated groundwater level.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

Characteristics of the groundwater level ('GHG' and 'GLG'; see also \link{get_gxg}) can be added to the plot by setting \code{add_gxg} to TRUE.

In case observations are available it is possible to add the observations as red dots in the plot.
A path to a folder with observations should be specified with \code{dir_obs}.
In this folder a csv-file with the name of the SWAP-variable should be present.
The file itself should have two columns: 'DATETIME' and 'VALUE'. The unit of the values should be equal to SWAP.
Optionally the model performance (\code{performance}) can be added in the plot (see also: \link{get_modelperformance}).

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, files = basename(c(file_swp, file_csv)), exdir = dir_out)

# plot timeseries of simulated groundwater level
plot_gwlevel(file_swp = file_swp, add_gxg = TRUE)

# clean example
unlink(x = dir_out, recursive = TRUE)
}
