% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_profile}
\alias{plot_profile}
\title{Create plot with selection of soil profiles of the Dutch soil map}
\usage{
plot_profile(...)
}
\arguments{
\item{...}{Arguments to pass, either \code{soil_id}, \code{profile_id}, \code{bofek_id} or \code{cluster_id} should be specified. Look at \code{Details} for further arguments to pass.}
}
\description{
Creating a plot for a selection of soil profiles (either specified by \code{soil_id}, \code{profile_id}, \code{bofek_id} or \code{cluster_id}).
}
\details{
Base information is the Dutch soil map (1:50,000) distinguishing 370 soil profiles (\href{https://edepot.wur.nl/298371}{de Vries, 1999}).
Each soil profile contains a sequence of soil layers that are linked to texture groups of the Staringreeks (\href{https://edepot.wur.nl/512761}{Heinen et al., 2020}).
Based on the soil physical properties, i.e. water retention and hydraulic conductivity, clusters of soil profiles can be created.

Soil profiles to be included in the plot can be selected by \code{soil_id} (options are 1-370) or \code{profile_id} (vector of soil profile numbers mentioned in \href{https://edepot.wur.nl/541544}{Heinen et al., 2021}).
Alternatively the dominant soil profiles within a BOFEK cluster can be selected by specifying \code{bofek_id}. All soil profiles within a BOFEK cluster can be selected by specifying a single \code{cluster_id}.
In case of a selection by \code{bofek_id} the type of BOFEK should be specified, options are 'BOFEK2012' (\href{https://edepot.wur.nl/247678}{Wosten et al., 2013}) or 'BOFEK2020' (\href{https://edepot.wur.nl/541544}{Heinen et al., 2021}).

The maximum depth of the soil profiles to be plotted is set to -120 cm below soil surface. Optionally this can be adjusted by specifying \code{max_depth}.

As default the soil profiles are arranged by area which is presented at the top of the graph, this can be turned of by setting \code{plot_area} to FALSE. In case \code{plot_area} is set to FALSE the soil profiles are arranged by user defined input.

The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.
}
\examples{
# plot selection of soil profiles
print(plot_profile(soil_id = c(36, 61, 79, 253, 367)))

# plot selection of bofek_id
print(plot_profile(bofek_id = c(1006, 2006, 3015, 4018, 5007), type = "BOFEK2020"))

# plot selection within BOFEK cluster
print(plot_profile(cluster_id = 3015, type = "BOFEK2020"))
}
