% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rain_irrig}
\alias{plot_rain_irrig}
\title{Create timeseries-plot of precipitation and irrigation}
\usage{
plot_rain_irrig(file_swp, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates timeseries-plot of precipitation and irrigation.
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) a timeseries-plot is created of the precipitation and irrigation.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

Optionally results can be aggregated in time by \code{aggregate} (options: "year", "month", "decade").

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
