% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_subreg}
\alias{plot_subreg}
\title{Create water balance plot for a subregion of the soil column}
\usage{
plot_subreg(file_swp, variable, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{variable}{character string of SWAP variable; options are 'SUBREG_MIN[\emph{range}]' or 'SUBREG_ALL[\emph{range}]'.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates water balance plot for a subregion of the soil column.
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) and SWAP variable (\code{variable}) a water balance plot is created for a subregion of the soil column.
The water balance can be created with net fluxes using variable 'SUBREG_MIN[\emph{range}]', or with all fluxes using variable 'SUBREG_ALL[\emph{range}]'.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

Optionally the terms of the water balance can be aggregated in time by \code{aggregate} (options: "year", "month", "decade").

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, files = basename(c(file_swp, file_csv)), exdir = dir_out)

# plot (netto) waterbalance of the first 40.0 cm from the top of the soil column
plot_subreg(file_swp = file_swp, variable = "SUBREG_MIN[0.0:-40.0]")

# plot (netto) waterbalance of the first 40.0 cm from the top of the soil column aggregated to monthly values
plot_subreg(file_swp = file_swp, variable = "SUBREG_MIN[0.0:-40.0]", aggregate = "month")

# plot waterbalance of the first 40.0 cm from the top of the soil column aggregated to monthly values
plot_subreg(file_swp = file_swp, variable = "SUBREG_ALL[0.0:-40.0]", aggregate = "month")

# clean example
unlink(x = dir_out, recursive = TRUE)
}
