% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tred}
\alias{plot_tred}
\title{Create timeseries-plot of simulated transpiration reduction}
\usage{
plot_tred(file_swp, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates timeseries-plot of transpiration reduction simulated by SWAP.
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) a timeseries-plot is created of the simulated transpiration reduction.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

Transpiration reduction can be caused by too dry ('drought'), too wet ('oxygen'), too saline ('salinity') or too cold ('frost') conditions.
By default all stressors are selected to present in the plot.
In case an stressor is not active (or not present in the model output) a list of stressors can be specified by \code{stressor}.

By default the transpiration reduction is plotted for all crops in the crop-calendar during the simulation period.
Specify a list of crops with \code{cropfil} in case only the transpiration reduction of specific crop(s) is preferred.

Optionally results can be aggregated in time by \code{aggregate} (options: "year", "month", "decade").

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, files = basename(c(file_swp, file_csv)), exdir = dir_out)

# plot timeseries of simulated transpiration reduction by drought stress
plot_tred(file_swp = file_swp, stressor = "drought")

# plot timeseries of simulated transpiration reduction by drought and oxygen stress
plot_tred(file_swp = file_swp, stressor = c("drought", "oxygen"))
# clean example
unlink(x = dir_out, recursive = TRUE)
}
