% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variable}
\alias{plot_variable}
\title{Create timeseries-plot}
\usage{
plot_variable(file_swp, variable, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{variable}{character string of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates timeseries-plot for a specific SWAP variable.
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) and SWAP variable (\code{variable}) a timeseries-plot is created.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

In case observations are available it is possible to add the observations as red dots in the plot.
A path to a folder with observations should be specified with \code{dir_obs}.
In this folder a csv-file with the name of the SWAP-variable should be present.
The file itself should have two columns: 'DATETIME' and 'VALUE'. The unit of the values should be equal to SWAP.
Optionally the model performance (\code{performance}) can be added in the plot (see also: \link{get_modelperformance}).

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, exdir = dir_out)

# plot timeseries of groundwater level
plot_variable(file_swp = file_swp, variable = "GWL")

# plot timeseries of groundwater level including observations and model performance
plot_variable(file_swp = file_swp, variable = "GWL", dir_obs = paste0(dir_out, "/observed"), performance = "rmse")

# plot timeseries of water content at 15 cm depth
plot_variable(file_swp = file_swp, variable = "WC[-15.0]")

# clean example
unlink(x = dir_out, recursive = TRUE)
}
