% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_yield}
\alias{plot_yield}
\title{Create timeseries-plot of potential and actual yield}
\usage{
plot_yield(file_swp, ...)
}
\arguments{
\item{file_swp}{character string of main SWAP inputfile (swp-file).}

\item{...}{further arguments passed to or from other methods}
}
\description{
Creates timeseries-plot of simulated potential and actual crop yield
}
\details{
Based on the main inputfile of SWAP (\code{file_swp}) a timeseries-plot is created of the potential and actual crop yield.
The plot can be adjusted by \code{tstart}, \code{tend} and \code{ylim}.
The labels of the plot can be set to 'EN' (English; default) or 'NL' (Dutch) using \code{language}.

By default the crop development of aboveground biomass is plotted for all crops (simulated by WOFOST; CROPTYPE = 2) in the crop-calendar during the simulation period.
Specify a list of crops with \code{cropfil} in case only the development of specific crop(s) is preferred.
In case the development of a certain part of the plant is preferred specify \code{growth} (options: aboveground biomass ('AGB') and storage organs ('SO').

The function will created an error by default which will stop the R-program. The error-message can be changed into a warning by setting \code{error} to 'FALSE'.
}
\examples{
# extract example
dir_out <- tempfile(pattern = "swap_")
file_swp <- paste0(dir_out, "/swap.swp")
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, exdir = dir_out)

# plot timeseries of crop development
plot_yield(file_swp = file_swp, growth = "AGB")

# clean example
unlink(x = dir_out, recursive = TRUE)
}
