% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{prepare_observed_SWAP}
\alias{prepare_observed_SWAP}
\title{Prepare observations belonging to SWAP simulation}
\usage{
prepare_observed_SWAP(file_swp, run_info)
}
\arguments{
\item{file_swp}{character string, name of swp-file.}

\item{run_info}{list, index of run_id.}
}
\description{
Prepare observations belonging to SWAP simulation
}
