% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{read_csv_SWAP}
\alias{read_csv_SWAP}
\title{Read SWAP-csv-output}
\usage{
read_csv_SWAP(file, variable, timestamp = FALSE, quiet = TRUE)
}
\arguments{
\item{file}{character string, name of SWAP csv-output-file.}

\item{variable}{character string, name of SWAP variable to extract.}

\item{timestamp}{logical, TRUE if multiple timesteps within a day.}

\item{quiet}{logical.}
}
\description{
Read SWAP-csv-output
}
\examples{
# extract example
dir_out <- paste0(tempfile(pattern = "swap_"))
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, files = basename(file_csv), exdir = dir_out)

# plot timeseries of groundwater level
read_csv_SWAP(file = file_csv, variable = c("GWL", "WC[-15.0]"))

# clean example
unlink(x = dir_out, recursive = TRUE)
}
