% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{read_SWAP}
\alias{read_SWAP}
\title{Read SWAP-output}
\usage{
read_SWAP(file, variable, timestamp = FALSE, quiet = TRUE)
}
\arguments{
\item{file}{character string, name of SWAP output-file.}

\item{variable}{character string, name of SWAP variable to extract.}

\item{timestamp}{logical, TRUE if multiple timesteps within a day.}

\item{quiet}{logical.}
}
\description{
Read SWAP-output
}
\keyword{internal}
