% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{run_SWAP}
\alias{run_SWAP}
\title{Run SWAP-model}
\usage{
run_SWAP(file_swp, command, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{file_swp}{character string, name of swp-file.}

\item{command}{character string, name of SWAP-program.}

\item{stop}{logical, execute R-script after SWAP crash.}

\item{quiet}{logical.}
}
\description{
Run SWAP-model
}
