% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{scan_csv_SWAP}
\alias{scan_csv_SWAP}
\title{Scan SWAP-csv-output}
\usage{
scan_csv_SWAP(file, filter = TRUE)
}
\arguments{
\item{file}{character string, name of SWAP csv-output-file.}

\item{filter}{logical, filter variables only.}
}
\description{
Scan SWAP-csv-output
}
\examples{
# extract example
dir_out <- paste0(tempfile(pattern = "swap_"))
file_csv <- paste0(dir_out, "/result_output.csv")
zipfile <- system.file("extdata/input.zip", package = "SWAPtools")
unzip(zipfile = zipfile, files = basename(file_csv), exdir = dir_out)

# get list of variables from output file of SWAP
scan_csv_SWAP(file = file_csv)

# clean example
unlink(x = dir_out, recursive = TRUE)
}
\keyword{internal}
