% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{set_switch_SWAP}
\alias{set_switch_SWAP}
\title{Set switch in SWAP-template}
\usage{
set_switch_SWAP(tmplt, switch, value, ...)
}
\arguments{
\item{tmplt}{character string, text of SWAP-template from which names of variable(s) should be extracted.}

\item{switch}{character string, name of switch in SWAP-template.}

\item{value}{integer, value of switch.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Set switch in SWAP-template
}
\keyword{internal}
