% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvG.R
\name{set_table_MvG}
\alias{set_table_MvG}
\title{Create MvG-table}
\usage{
set_table_MvG(file, WCR, WCS, ALPHA, NPAR, LAMBDA, KSAT)
}
\arguments{
\item{file}{filename.}

\item{WCR}{residual water content [cm3 cm-3]}

\item{WCS}{saturated water content [cm3 cm-3]}

\item{ALPHA}{curve shape parameter [-]}

\item{NPAR}{curve shape parameter [-]}

\item{LAMBDA}{curve shape parameter [-]}

\item{KSAT}{hydraulic conductivity of saturated soil [cm d-1]}
}
\description{
Create MvG-table
}
