% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{set_table_SWAP}
\alias{set_table_SWAP}
\title{Set table in SWAP-template}
\usage{
set_table_SWAP(tmplt, table, value, ...)
}
\arguments{
\item{tmplt}{character string, text of SWAP-template from which names of variable(s) should be extracted.}

\item{table}{character string, name of table in SWAP-template from which names of variables should be extracted.}

\item{value}{table, value of variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Set table in SWAP-template
}
\keyword{internal}
