% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{set_template_SWAP}
\alias{set_template_SWAP}
\title{Set SWAP-template}
\usage{
set_template_SWAP(template, ...)
}
\arguments{
\item{template}{character string, name of SWAP-template.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Set SWAP-template
}
\keyword{internal}
