% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{set_value_SWAP}
\alias{set_value_SWAP}
\title{Set value of SWAP-variable}
\usage{
set_value_SWAP(value, variable, ...)
}
\arguments{
\item{value}{Value (numeric, date or character string) of SWAP-variable.}

\item{variable}{character string, name of SWAP variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Set value of SWAP-variable
}
\keyword{internal}
