% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{set_variable_SWAP}
\alias{set_variable_SWAP}
\title{Set variable(s) in SWAP-template}
\usage{
set_variable_SWAP(tmplt, variable, value, ...)
}
\arguments{
\item{tmplt}{character string, text of SWAP-template from which names of variable(s) should be extracted.}

\item{variable}{character string, name of SWAP variable.}

\item{value}{numeric, date or string, value of variable.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Set variable(s) in SWAP-template
}
\keyword{internal}
