% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{summarise_db}
\alias{summarise_db}
\title{Aggregate data in time}
\usage{
summarise_db(db, level, ...)
}
\arguments{
\item{db}{tibble to aggregate.}

\item{level}{level of aggregation.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Aggregate data in time
}
\details{
Summarise data in time set by \code{level} (options: "year", "month", "decade").
Optionally columns to summarise can be set by \code{column}.
}
\keyword{internal}
