% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\name{zip_SWAP}
\alias{zip_SWAP}
\title{Zip SWAP-model results}
\usage{
zip_SWAP(file_swp, glob = NULL)
}
\arguments{
\item{file_swp}{character string, name of swp-file.}

\item{glob}{pattern to filter files, optional.}
}
\description{
Zip SWAP-model results
}
